+++ .classpath
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="lib" path="build/libs/CameraVision-all.jar">
+		<attributes>
+			<attribute name="org.eclipse.jdt.launching.CLASSPATH_ATTR_LIBRARY_PATH_ENTRY" value="TrackerboxReloaded/build/opencv"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="build/libs/CameraVision.jar"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
