+++ build.gradle
@@ -10,16 +10,16 @@ apply plugin: 'application'
 // This is where you select which operating system to build for.
 
 // Uncomment the line below to build for windows
-ext.buildType = "windows"
+//ext.buildType = "windows"
 
 // Uncomment the line below to build for a Raspberry Pi running raspbian
-//ext.buildType = "arm-raspbian"
+ext.buildType = "arm-raspbian"
 
 // Uncomment the line below to build for an armhf device such as a Jetson or a Beaglebone Black
 //ext.buildType = "armhf"
 
 // Change the line below if you change the name of your main Java class
-mainClassName = 'Main'
+mainClassName = 'ca.team2706.vision.trackerboxreloaded.Main'
 // Change the line below to change the name of the output jar
 def projectName = 'CameraVision'
