+++ output/runCameraVision.bat
@@ -1,0 +1,1 @@
+java -Djava.library.path=. -jar CameraVision-all.jar
